fix: fix HttpExceptionsFilter
