+++ src/main.ts
@@ -1,4 +1,4 @@
-import { HttpAdapterHost, NestFactory } from '@nestjs/core';
+import { NestFactory } from '@nestjs/core';
 import { env } from 'process';
 import { AppModule } from './app.module';
 import * as dotenv from 'dotenv';
@@ -7,7 +7,7 @@ import { OpenAPIObject, SwaggerModule } from '@nestjs/swagger';
 import { readFileSync } from 'fs';
 import { join } from 'path';
 import { LoggerService } from './logger/logger.service';
-import { LogExceptionsFilter } from './logger/log-exception.filter';
+import { HttpExceptionsFilter } from './exceptionsFilter/exception.filter';
 
 dotenv.config();
 
@@ -21,7 +21,7 @@ async function bootstrap() {
   const logger = new LoggerService();
 
   app.useLogger(logger);
-  app.useGlobalFilters(new LogExceptionsFilter(new HttpAdapterHost(), logger));
+  app.useGlobalFilters(new HttpExceptionsFilter(logger));
 
   const document = load(
     readFileSync(join(__dirname, '..', 'doc', 'doc.yaml'), 'utf8'),
